+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/scenarios/InsertDepartment.java
@@ -21,7 +21,7 @@ package is.codion.framework.demos.employees.testing.scenarios;
 import is.codion.common.model.loadtest.LoadTest.Scenario.Performer;
 import is.codion.framework.demos.employees.domain.Employees.Department;
 import is.codion.framework.demos.employees.model.EmployeesAppModel;
-import is.codion.framework.domain.entity.test.DefaultEntityFactory;
+import is.codion.framework.domain.test.DefaultEntityFactory;
 import is.codion.swing.framework.model.SwingEntityModel;
 
 // tag::loadTest[]